requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pillow==10.4.0
numpy==1.26.4
google-genai==0.8.0
jinja2==3.1.4
click==8.1.7
python-multipart==0.0.9
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
requests==2.32.3


